everything working, going to try and use prettier to make formatting nicer
