+++ src/Components/CodeContainer.js
@@ -7,7 +7,7 @@ class CodeContainer extends Component {
     }
 
     renderSecret(){
-        if(this.props.revealCode){
+        if(this.props.revealCode || this.props.tries === 0){
             return this.props.secret.map((number, index) => <img key={index} alt={number} src={`/img/chars/${number}.jpg`}></img>)
         } else {
             return this.props.secret.map((number, index) => <img key={index} alt={number} src="/img/mystery/block.jpg"></img>)
