+++ src/Components/StartForm.js
@@ -24,48 +24,38 @@ class StartForm extends Component {
 
     render(){
         return (
-            <form className='start-form' onSubmit={this.localStartHandler}>
-                <h4>Time Limit</h4>
-                <label for='None'>None</label>
-                <input type='radio' name='timer' value={0} onChange={this.changeHandler}/>
-                <label for='30'>30s</label>
-                <input type='radio' name='timer' value={30} onChange={this.changeHandler}/>
-                <label for='15'>15s</label>
-                <input type='radio' name='timer' value={15} onChange={this.changeHandler}/>
-                <h4>Amount of Options</h4>
-                <label for='6'>6</label>
-                <input type='radio' name='options' value={6} onChange={this.changeHandler}/>
-                <label for='8'>8</label>
-                <input type='radio' name='options' value={8} onChange={this.changeHandler}/>
-                <label for='10'>10</label>
-                <input type='radio' name='options' value={10} onChange={this.changeHandler}/>
-            <button type='submit'>Start Game</button>  
-        </form>
+            <div>
+                <h3 className='rules-header'>Rules of Mastermind</h3>
+                <p>When you press start, a secret code of 4 characters will be generated by the computer.</p>
+                <p>The code is completely random, meaning that the same character can be repeated.</p>
+                <p>You will have 10 attempts to guess the correct code.</p>
+                <p>After each guess, you will be given feedback.</p>
+                <p>For a character in the right place, you get a star!</p>
+                <p>For a character that is included but in the wrong place, you get a coin.</p>
+                <p>If the character is not in the secret code, you get a banana.</p>
+                <h3 className='rules-header'>Settings</h3>
+                <p>If you choose to add a timer, whenever it runs out, Bowser will make a guess for you.</p>
+                <p>You can also adjust the number of possible characters in the code. Think of 6 as easy, 8 as normal, and 10 as hard.</p>
+                <form className='start-form' onSubmit={this.localStartHandler}>
+                    <h4>Time Limit</h4>
+                    <label for='None'>None</label>
+                    <input type='radio' name='timer' value={0} onChange={this.changeHandler}/>
+                    <label for='30'>30s</label>
+                    <input type='radio' name='timer' value={30} onChange={this.changeHandler}/>
+                    <label for='15'>15s</label>
+                    <input type='radio' name='timer' value={15} onChange={this.changeHandler}/>
+                    <h4>Amount of Options</h4>
+                    <label for='6'>6</label>
+                    <input type='radio' name='options' value={6} onChange={this.changeHandler}/>
+                    <label for='8'>8</label>
+                    <input type='radio' name='options' value={8} onChange={this.changeHandler}/>
+                    <label for='10'>10</label>
+                    <input type='radio' name='options' value={10} onChange={this.changeHandler}/>
+                    <button type='submit'>Start Game</button>  
+                </form>
+            </div>
         )
     }
 }
 
 export default StartForm
-
-
-
-
-// return (
-//     <form className='start-form' onSubmit={this.localStartHandler}>
-//         <h4>Time Limit</h4>
-//         <label for='None'>None</label>
-//         <input type='radio' name='time' value='0' />
-//         <label for='30'>30s</label>
-//         <input type='radio' name='time' value='30' />
-//         <label for='15'>15s</label>
-//         <input type='radio' name='time' value='15' />
-//         <h4>Amount of Options</h4>
-//         <label for='6'>6</label>
-//         <input type='radio' name='options' value='6' />
-//         <label for='30'>8</label>
-//         <input type='radio' name='options' value='8' />
-//         <label for='15'>10</label>
-//         <input type='radio' name='options' value='10' />
-//     <button type='submit'>Start Game</button>  
-// </form>
-// )
